+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.7.0] - 2021-02-22
+### Added
+- Notice if The SEO Framework options have not been reseted to our standards
+
+### Fixed
+- Hiding some The SEO Framework features
+
 ## [2.6.0] - 2021-02-19
 ### Changed
 - Allow REST API users endpoints if user is logged in and can edit_posts